Tried implementing what I think part 5 is
My thought process is that I should use 8 to 1 multiplexor, with the 3 inputs as each bit of from ALUControl, and somehow use that in a 1 bit ALU function. I repeat that 32 times, but I do not have any idea on what to put the 8-to-1 multiplexor excluding the three select lines.  Any help?? I'm sorry I have been looking for so long but I can't seem to find an answer. I also tried implementing Extend_Sign16 function